Conflicts:
	modules/perception/traffic_light/base/image.cc
	modules/perception/traffic_light/base/image.h
	modules/perception/traffic_light/base/image_lights.h
	modules/perception/traffic_light/base/light.cc
	modules/perception/traffic_light/base/light.h
	modules/perception/traffic_light/base/test/image_test.cc
	modules/perception/traffic_light/base/test/light_test.cc
	modules/perception/traffic_light/base/test/utils_structs_test.cc
	modules/perception/traffic_light/base/tl_shared_data.cc
	modules/perception/traffic_light/base/utils.cc
	modules/perception/traffic_light/base/utils.h
	modules/perception/traffic_light/interface/base_projection.cc
	modules/perception/traffic_light/interface/base_projection.h
	modules/perception/traffic_light/interface/base_rectifier.h
	modules/perception/traffic_light/interface/base_reviser.h
	modules/perception/traffic_light/onboard/hdmap_input.cc
	modules/perception/traffic_light/onboard/hdmap_input.h
	modules/perception/traffic_light/onboard/preprocessor_subnode.cc
	modules/perception/traffic_light/onboard/preprocessor_subnode.h
	modules/perception/traffic_light/onboard/proc_subnode.cc
	modules/perception/traffic_light/onboard/proc_subnode.h
	modules/perception/traffic_light/preprocessor/test/tl_preprocessor_test.cc
	modules/perception/traffic_light/preprocessor/tl_preprocessor.cc
	modules/perception/traffic_light/preprocessor/tl_preprocessor.h
	modules/perception/traffic_light/projection/multi_camera_projection.cc
	modules/perception/traffic_light/projection/multi_camera_projection.h
	modules/perception/traffic_light/projection/projection.cc
	modules/perception/traffic_light/projection/projection.h
	modules/perception/traffic_light/recognizer/classify.cc
	modules/perception/traffic_light/recognizer/classify.h
	modules/perception/traffic_light/recognizer/unity_recognize.cc
	modules/perception/traffic_light/recognizer/unity_recognize.h
	modules/perception/traffic_light/rectify/cropbox.cc
	modules/perception/traffic_light/rectify/cropbox.h
	modules/perception/traffic_light/rectify/detection.cc
	modules/perception/traffic_light/rectify/detection.h
	modules/perception/traffic_light/rectify/munkres.cc
	modules/perception/traffic_light/rectify/munkres.h
	modules/perception/traffic_light/rectify/select.cc
	modules/perception/traffic_light/rectify/select.h
	modules/perception/traffic_light/rectify/unity_rectify.cc
	modules/perception/traffic_light/rectify/unity_rectify.h
	modules/perception/traffic_light/reviser/color_decision.cc
	modules/perception/traffic_light/reviser/color_decision.h
	modules/perception/traffic_light/util/color_space.cc
	modules/perception/traffic_light/util/color_space.h

+++ modules/perception/traffic_light/base/image_lights.h
@@ -17,8 +17,8 @@
 #ifndef MODULES_PERCEPTION_TRAFFIC_LIGHT_ONBOARD_IMAGE_LIGHTS_H
 #define MODULES_PERCEPTION_TRAFFIC_LIGHT_ONBOARD_IMAGE_LIGHTS_H
 
-#include <memory>
 #include <vector>
+#include <memory>
 
 #include "modules/perception/traffic_light/base/image.h"
 #include "modules/perception/traffic_light/base/light.h"
@@ -32,19 +32,21 @@ struct ImageLights {
   std::shared_ptr<Image> image;
   CarPose pose;
   std::shared_ptr<LightPtrs> lights;
-  std::shared_ptr<LightPtrs>
-      lights_outside_image;  // record the lights outside the lights.
+  // record the lights outside the lights.
+  std::shared_ptr<LightPtrs> lights_outside_image;
   CameraId camera_id = UNKNOWN;
-  double timestamp = 0.0;  // image's timestamp
-
-  double preprocess_receive_timestamp = 0.0;  // timestamp when received a image
-  double preprocess_send_timestamp =
-      0.0;  // timestamp when PreprocessSubnode pub event
+  double timestamp = 0.0;   // image's timestamp
+  // timestamp when received a image
+  double preprocess_receive_timestamp = 0.0;
+  // timestamp when PreprocessSubnode pub event
+  double preprocess_send_timestamp = 0.0;
   bool is_pose_valid = false;
-  double diff_image_pose_ts =
-      0.0;  // image' timestamp minus the most recently pose's timestamp
-  double diff_image_sys_ts = 0.0;  // image' timestamp system's timestamp
-  size_t offset = 0;  // offset size between hdmap bbox and detection bbox
+  // image' timestamp minus the most recently pose's timestamp
+  double diff_image_pose_ts = 0.0;
+  // image' timestamp system's timestamp
+  double diff_image_sys_ts = 0.0;
+  // offset size between hdmap bbox and detection bbox
+  int offset = 0;
   size_t num_signals = 0;
 };
